// src/app/shared/interfaces.ts
export interface User {
  email: string
  password: string
  returnSecureToken: boolean
}

export interface FbAuthResponse {
  idToken: string
  expiresIn: string
}

export interface Post {
  id?: string
  title: string
  text: string
  author: string
  date: Date
}

export interface PostsResponse {
  [key: string]: Post
}

export interface FbCreateResponse {
  name: string
}